requirements.txt
requests
newspaper3k
python-dotenv
openai
